esm - convert window related helpers to TS
