--- src/bootstrap-window.js
@@ -1,257 +1,0 @@
-/*---------------------------------------------------------------------------------------------
- *  Copyright (c) Microsoft Corporation. All rights reserved.
- *  Licensed under the MIT License. See License.txt in the project root for license information.
- *--------------------------------------------------------------------------------------------*/
-
-//@ts-check
-'use strict';
-
-/**
- * @import { ISandboxConfiguration } from './vs/base/parts/sandbox/common/sandboxTypes'
- * @typedef {any} LoaderConfig
- */
-
-/* eslint-disable no-restricted-globals */
-
-(function (factory) {
-	// @ts-ignore
-	globalThis.MonacoBootstrapWindow = factory();
-}(function () {
-	const preloadGlobals = sandboxGlobals();
-	const safeProcess = preloadGlobals.process;
-
-	// increase number of stack frames(from 10, https://github.com/v8/v8/wiki/Stack-Trace-API)
-	Error.stackTraceLimit = 100;
-
-	/**
-	 * @param {string} esModule
-	 * @param {(result: unknown, configuration: ISandboxConfiguration) => Promise<unknown> | undefined} resultCallback
-	 * @param {{
-	 *  configureDeveloperSettings?: (config: ISandboxConfiguration) => {
-	 * 		forceDisableShowDevtoolsOnError?: boolean,
-	 * 		forceEnableDeveloperKeybindings?: boolean,
-	 * 		disallowReloadKeybinding?: boolean,
-	 * 		removeDeveloperKeybindingsAfterLoad?: boolean
-	 * 	},
-	 * 	canModifyDOM?: (config: ISandboxConfiguration) => void,
-	 *  beforeImport?: (config: ISandboxConfiguration) => void
-	 * }} [options]
-	 */
-	async function load(esModule, resultCallback, options) {
-
-		// Await window configuration from preload
-		const timeout = setTimeout(() => { console.error(`[resolve window config] Could not resolve window configuration within 10 seconds, but will continue to wait...`); }, 10000);
-		performance.mark('code/willWaitForWindowConfig');
-		/** @type {ISandboxConfiguration} */
-		const configuration = await preloadGlobals.context.resolveConfiguration();
-		performance.mark('code/didWaitForWindowConfig');
-		clearTimeout(timeout);
-
-		// Signal DOM modifications are now OK
-		if (typeof options?.canModifyDOM === 'function') {
-			options.canModifyDOM(configuration);
-		}
-
-		// Developer settings
-		const {
-			forceEnableDeveloperKeybindings,
-			disallowReloadKeybinding,
-			removeDeveloperKeybindingsAfterLoad
-		} = typeof options?.configureDeveloperSettings === 'function' ? options.configureDeveloperSettings(configuration) : {
-			forceEnableDeveloperKeybindings: false,
-			disallowReloadKeybinding: false,
-			removeDeveloperKeybindingsAfterLoad: false
-		};
-		const isDev = !!safeProcess.env['VSCODE_DEV'];
-		const enableDeveloperKeybindings = isDev || forceEnableDeveloperKeybindings;
-		let developerDeveloperKeybindingsDisposable = undefined;
-		if (enableDeveloperKeybindings) {
-			developerDeveloperKeybindingsDisposable = registerDeveloperKeybindings(disallowReloadKeybinding);
-		}
-
-		globalThis._VSCODE_NLS_MESSAGES = configuration.nls.messages;
-		globalThis._VSCODE_NLS_LANGUAGE = configuration.nls.language;
-		let language = configuration.nls.language || 'en';
-		if (language === 'zh-tw') {
-			language = 'zh-Hant';
-		} else if (language === 'zh-cn') {
-			language = 'zh-Hans';
-		}
-
-		window.document.documentElement.setAttribute('lang', language);
-
-		window['MonacoEnvironment'] = {};
-
-		// Signal before import()
-		if (typeof options?.beforeImport === 'function') {
-			options.beforeImport(configuration);
-		}
-
-		const baseUrl = new URL(`${fileUriFromPath(configuration.appRoot, { isWindows: safeProcess.platform === 'win32', scheme: 'vscode-file', fallbackAuthority: 'vscode-app' })}/out/`);
-		globalThis._VSCODE_FILE_ROOT = baseUrl.toString();
-
-		// DEV ---------------------------------------------------------------------------------------
-		// DEV: This is for development and enables loading CSS via import-statements via import-maps.
-		// DEV: For each CSS modules that we have we defined an entry in the import map that maps to
-		// DEV: a blob URL that loads the CSS via a dynamic @import-rule.
-		// DEV ---------------------------------------------------------------------------------------
-		if (Array.isArray(configuration.cssModules) && configuration.cssModules.length > 0) {
-			performance.mark('code/willAddCssLoader');
-
-			const style = document.createElement('style');
-			style.type = 'text/css';
-			style.media = 'screen';
-			style.id = 'vscode-css-loading';
-			document.head.appendChild(style);
-
-			globalThis._VSCODE_CSS_LOAD = function (url) {
-				style.textContent += `@import url(${url});\n`;
-			};
-
-			/**
-			 * @type { { imports: Record<string, string> }}
-			 */
-			const importMap = { imports: {} };
-			for (const cssModule of configuration.cssModules) {
-				const cssUrl = new URL(cssModule, baseUrl).href;
-				const jsSrc = `globalThis._VSCODE_CSS_LOAD('${cssUrl}');\n`;
-				const blob = new Blob([jsSrc], { type: 'application/javascript' });
-				importMap.imports[cssUrl] = URL.createObjectURL(blob);
-			}
-
-			const ttp = window.trustedTypes?.createPolicy('vscode-bootstrapImportMap', { createScript(value) { return value; }, });
-			const importMapSrc = JSON.stringify(importMap, undefined, 2);
-			const importMapScript = document.createElement('script');
-			importMapScript.type = 'importmap';
-			importMapScript.setAttribute('nonce', '0c6a828f1297');
-			// @ts-ignore
-			importMapScript.textContent = ttp?.createScript(importMapSrc) ?? importMapSrc;
-			document.head.appendChild(importMapScript);
-
-			performance.mark('code/didAddCssLoader');
-		}
-
-		// ESM Import
-		try {
-			const result = await import(new URL(`${esModule}.js`, baseUrl).href);
-
-			const callbackResult = resultCallback(result, configuration);
-			if (callbackResult instanceof Promise) {
-				await callbackResult;
-
-				if (developerDeveloperKeybindingsDisposable && removeDeveloperKeybindingsAfterLoad) {
-					developerDeveloperKeybindingsDisposable();
-				}
-			}
-		} catch (error) {
-			onUnexpectedError(error, enableDeveloperKeybindings);
-		}
-	}
-
-	/**
-	 * @param {boolean | undefined} disallowReloadKeybinding
-	 * @returns {() => void}
-	 */
-	function registerDeveloperKeybindings(disallowReloadKeybinding) {
-		const ipcRenderer = preloadGlobals.ipcRenderer;
-
-		const extractKey =
-			/**
-			 * @param {KeyboardEvent} e
-			 */
-			function (e) {
-				return [
-					e.ctrlKey ? 'ctrl-' : '',
-					e.metaKey ? 'meta-' : '',
-					e.altKey ? 'alt-' : '',
-					e.shiftKey ? 'shift-' : '',
-					e.keyCode
-				].join('');
-			};
-
-		// Devtools & reload support
-		const TOGGLE_DEV_TOOLS_KB = (safeProcess.platform === 'darwin' ? 'meta-alt-73' : 'ctrl-shift-73'); // mac: Cmd-Alt-I, rest: Ctrl-Shift-I
-		const TOGGLE_DEV_TOOLS_KB_ALT = '123'; // F12
-		const RELOAD_KB = (safeProcess.platform === 'darwin' ? 'meta-82' : 'ctrl-82'); // mac: Cmd-R, rest: Ctrl-R
-
-		/** @type {((e: KeyboardEvent) => void) | undefined} */
-		let listener = function (e) {
-			const key = extractKey(e);
-			if (key === TOGGLE_DEV_TOOLS_KB || key === TOGGLE_DEV_TOOLS_KB_ALT) {
-				ipcRenderer.send('vscode:toggleDevTools');
-			} else if (key === RELOAD_KB && !disallowReloadKeybinding) {
-				ipcRenderer.send('vscode:reloadWindow');
-			}
-		};
-
-		window.addEventListener('keydown', listener);
-
-		return function () {
-			if (listener) {
-				window.removeEventListener('keydown', listener);
-				listener = undefined;
-			}
-		};
-	}
-
-	/**
-	 * @param {string | Error} error
-	 * @param {boolean} [showDevtoolsOnError]
-	 */
-	function onUnexpectedError(error, showDevtoolsOnError) {
-		if (showDevtoolsOnError) {
-			const ipcRenderer = preloadGlobals.ipcRenderer;
-			ipcRenderer.send('vscode:openDevTools');
-		}
-
-		console.error(`[uncaught exception]: ${error}`);
-
-		if (error && typeof error !== 'string' && error.stack) {
-			console.error(error.stack);
-		}
-	}
-
-	/**
-	 * @param {string} path
-	 * @param {{ isWindows?: boolean, scheme?: string, fallbackAuthority?: string }} config
-	 * @returns {string}
-	 */
-	function fileUriFromPath(path, config) {
-
-		// Since we are building a URI, we normalize any backslash
-		// to slashes and we ensure that the path begins with a '/'.
-		let pathName = path.replace(/\\/g, '/');
-		if (pathName.length > 0 && pathName.charAt(0) !== '/') {
-			pathName = `/${pathName}`;
-		}
-
-		/** @type {string} */
-		let uri;
-
-		// Windows: in order to support UNC paths (which start with '//')
-		// that have their own authority, we do not use the provided authority
-		// but rather preserve it.
-		if (config.isWindows && pathName.startsWith('//')) {
-			uri = encodeURI(`${config.scheme || 'file'}:${pathName}`);
-		}
-
-		// Otherwise we optionally add the provided authority if specified
-		else {
-			uri = encodeURI(`${config.scheme || 'file'}://${config.fallbackAuthority || ''}${pathName}`);
-		}
-
-		return uri.replace(/#/g, '%23');
-	}
-
-	/**
-	 * @return {typeof import('./vs/base/parts/sandbox/electron-sandbox/globals')}
-	 */
-	function sandboxGlobals() {
-		// @ts-ignore (defined in globals.js)
-		return window.vscode;
-	}
-
-	return {
-		load
-	};
-}));
