+++ src/tsconfig.json
@@ -33,7 +33,7 @@
 		"./bootstrap-meta.ts",
 		"./bootstrap-node.ts",
 		"./bootstrap-server.ts",
-		"./bootstrap-window.js",
+		"./bootstrap-window.ts",
 		"./cli.ts",
 		"./main.ts",
 		"./server-main.ts",
@@ -45,9 +45,9 @@
 		"./vs/platform/environment/node/userDataPath.ts",
 		"./vs/base/parts/sandbox/electron-sandbox/preload-aux.ts",
 		"./vs/base/parts/sandbox/electron-sandbox/preload.ts",
-		"./vs/code/electron-sandbox/processExplorer/processExplorer.js",
-		"./vs/code/electron-sandbox/workbench/workbench.js",
-		"./vs/workbench/contrib/issue/electron-sandbox/issueReporter.js",
+		"./vs/code/electron-sandbox/processExplorer/processExplorer.ts",
+		"./vs/code/electron-sandbox/workbench/workbench.ts",
+		"./vs/workbench/contrib/issue/electron-sandbox/issueReporter.ts",
 		"./typings",
 		"./vs/**/*.ts",
 		"vscode-dts/vscode.proposed.*.d.ts",
