+++ src/vs/base/parts/sandbox/electron-sandbox/preload.ts
@@ -5,11 +5,11 @@
 
 /* eslint-disable no-restricted-globals */
 
-type SandboxConfiguration = import('vs/base/parts/sandbox/common/sandboxTypes').ISandboxConfiguration;
-
 (function () {
 
 	const { ipcRenderer, webFrame, contextBridge, webUtils } = require('electron');
+
+	type ISandboxConfiguration = import('vs/base/parts/sandbox/common/sandboxTypes.js').ISandboxConfiguration;
 
 	//#region Utilities
 
@@ -35,9 +35,9 @@ type SandboxConfiguration = import('vs/base/parts/sandbox/common/sandboxTypes').
 
 	//#region Resolve Configuration
 
-	let configuration: SandboxConfiguration | undefined = undefined;
+	let configuration: ISandboxConfiguration | undefined = undefined;
 
-	const resolveConfiguration: Promise<SandboxConfiguration> = (async () => {
+	const resolveConfiguration: Promise<ISandboxConfiguration> = (async () => {
 		const windowConfigIpcChannel = parseArgv('vscode-window-config');
 		if (!windowConfigIpcChannel) {
 			throw new Error('Preload: did not find expected vscode-window-config in renderer process arguments list.');
@@ -47,7 +47,7 @@ type SandboxConfiguration = import('vs/base/parts/sandbox/common/sandboxTypes').
 			validateIPC(windowConfigIpcChannel);
 
 			// Resolve configuration from electron-main
-			const resolvedConfiguration: SandboxConfiguration = configuration = await ipcRenderer.invoke(windowConfigIpcChannel);
+			const resolvedConfiguration: ISandboxConfiguration = configuration = await ipcRenderer.invoke(windowConfigIpcChannel);
 
 			// Apply `userEnv` directly
 			Object.assign(process.env, resolvedConfiguration.userEnv);
@@ -234,14 +234,14 @@ type SandboxConfiguration = import('vs/base/parts/sandbox/common/sandboxTypes').
 			 * actual value will be set after `resolveConfiguration`
 			 * has finished.
 			 */
-			configuration(): SandboxConfiguration | undefined {
+			configuration(): ISandboxConfiguration | undefined {
 				return configuration;
 			},
 
 			/**
 			 * Allows to await the resolution of the configuration object.
 			 */
-			async resolveConfiguration(): Promise<SandboxConfiguration> {
+			async resolveConfiguration(): Promise<ISandboxConfiguration> {
 				return resolveConfiguration;
 			}
 		}
