+++ src/vs/code/electron-sandbox/workbench/workbench.ts
@@ -3,18 +3,19 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-//@ts-check
-'use strict';
+/* eslint-disable no-restricted-globals */
 
 (function () {
 
-	/**
-	 * @import {INativeWindowConfiguration} from '../../../platform/window/common/window'
-	 * @import {NativeParsedArgs} from '../../../platform/environment/common/argv'
-	 * @import {ISandboxConfiguration} from '../../../base/parts/sandbox/common/sandboxTypes'
-	 */
+	type INativeWindowConfiguration = import('vs/platform/window/common/window.ts').INativeWindowConfiguration;
+	type NativeParsedArgs = import('vs/platform/environment/common/argv.js').NativeParsedArgs;
+	type IBootstrapWindow = import('vs/platform/window/electron-sandbox/window.js').IBootstrapWindow;
+	type PreloadGlobals = typeof import('vs/base/parts/sandbox/electron-sandbox/globals.js');
 
-	const bootstrapWindow = bootstrapWindowLib();
+	// @ts-ignore (defined in bootstrap-window.js)
+	const bootstrapWindow: IBootstrapWindow = window.MonacoBootstrapWindow;
+	// @ts-ignore (defined in preload.ts)
+	const preloadGlobals: PreloadGlobals = window.vscode;
 
 	// Add a perf entry right from the top
 	performance.mark('code/didStartRenderer');
@@ -23,7 +24,7 @@
 	// optimization to prevent a waterfall of loading to happen, because
 	// we know for a fact that workbench.desktop.main will depend on
 	// the related CSS counterpart.
-	bootstrapWindow.load('vs/workbench/workbench.desktop.main',
+	bootstrapWindow.load<INativeWindowConfiguration>('vs/workbench/workbench.desktop.main',
 		function (desktopMain, configuration) {
 
 			// Mark start of workbench
@@ -71,33 +72,7 @@
 
 	//#region Helpers
 
-	/**
-	 * @returns {{
-	 *   load: (
-	 *     esModule: string,
-	 *     resultCallback: (result: any, configuration: INativeWindowConfiguration & NativeParsedArgs) => unknown,
-	 *     options?: {
-	 *       configureDeveloperSettings?: (config: INativeWindowConfiguration & NativeParsedArgs) => {
-	 * 			forceDisableShowDevtoolsOnError?: boolean,
-	 * 			forceEnableDeveloperKeybindings?: boolean,
-	 * 			disallowReloadKeybinding?: boolean,
-	 * 			removeDeveloperKeybindingsAfterLoad?: boolean
-	 * 		 },
-	 * 	     canModifyDOM?: (config: INativeWindowConfiguration & NativeParsedArgs) => void,
-	 *       beforeImport?: (config: ISandboxConfiguration) => void
-	 *     }
-	 *   ) => Promise<unknown>
-	 * }}
-	 */
-	function bootstrapWindowLib() {
-		// @ts-ignore (defined in bootstrap-window.js)
-		return window.MonacoBootstrapWindow;
-	}
-
-	/**
-	 * @param {INativeWindowConfiguration & NativeParsedArgs} configuration
-	 */
-	function showSplash(configuration) {
+	function showSplash(configuration: INativeWindowConfiguration & NativeParsedArgs) {
 		performance.mark('code/willShowPartsSplash');
 
 		let data = configuration.partsSplash;
@@ -153,19 +128,17 @@
 
 		const style = document.createElement('style');
 		style.className = 'initialShellColors';
-		document.head.appendChild(style);
+		window.document.head.appendChild(style);
 		style.textContent = `body {
-			background-color: ${shellBackground};
-			color: ${shellForeground};
-			margin: 0;
-			padding: 0;
-		}`;
+				background-color: ${shellBackground};
+				color: ${shellForeground};
+				margin: 0;
+				padding: 0;
+			}`;
 
 		// set zoom level as soon as possible
-		// @ts-ignore
-		if (typeof data?.zoomLevel === 'number' && typeof globalThis.vscode?.webFrame?.setZoomLevel === 'function') {
-			// @ts-ignore
-			globalThis.vscode.webFrame.setZoomLevel(data.zoomLevel);
+		if (typeof data?.zoomLevel === 'number' && typeof preloadGlobals?.webFrame?.setZoomLevel === 'function') {
+			preloadGlobals.webFrame.setZoomLevel(data.zoomLevel);
 		}
 
 		// restore parts if possible (we might not always store layout info)
@@ -178,11 +151,11 @@
 
 			if (layoutInfo.windowBorder && colorInfo.windowBorder) {
 				splash.setAttribute('style', `
-					position: relative;
-					height: calc(100vh - 2px);
-					width: calc(100vw - 2px);
-					border: 1px solid var(--window-border-color);
-				`);
+						position: relative;
+						height: calc(100vh - 2px);
+						width: calc(100vw - 2px);
+						border: 1px solid var(--window-border-color);
+					`);
 				splash.style.setProperty('--window-border-color', colorInfo.windowBorder);
 
 				if (layoutInfo.windowBorderRadius) {
@@ -196,51 +169,51 @@
 			// part: title
 			const titleDiv = document.createElement('div');
 			titleDiv.setAttribute('style', `
-				position: absolute;
-				width: 100%;
-				height: ${layoutInfo.titleBarHeight}px;
-				left: 0;
-				top: 0;
-				background-color: ${colorInfo.titleBarBackground};
-				-webkit-app-region: drag;
-			`);
+					position: absolute;
+					width: 100%;
+					height: ${layoutInfo.titleBarHeight}px;
+					left: 0;
+					top: 0;
+					background-color: ${colorInfo.titleBarBackground};
+					-webkit-app-region: drag;
+				`);
 			splash.appendChild(titleDiv);
 
 			if (colorInfo.titleBarBorder && layoutInfo.titleBarHeight > 0) {
 				const titleBorder = document.createElement('div');
 				titleBorder.setAttribute('style', `
-					position: absolute;
-					width: 100%;
-					height: 1px;
-					left: 0;
-					bottom: 0;
-					border-bottom: 1px solid ${colorInfo.titleBarBorder};
-				`);
+						position: absolute;
+						width: 100%;
+						height: 1px;
+						left: 0;
+						bottom: 0;
+						border-bottom: 1px solid ${colorInfo.titleBarBorder};
+					`);
 				titleDiv.appendChild(titleBorder);
 			}
 
 			// part: activity bar
 			const activityDiv = document.createElement('div');
 			activityDiv.setAttribute('style', `
-				position: absolute;
-				width: ${layoutInfo.activityBarWidth}px;
-				height: calc(100% - ${layoutInfo.titleBarHeight + layoutInfo.statusBarHeight}px);
-				top: ${layoutInfo.titleBarHeight}px;
-				${layoutInfo.sideBarSide}: 0;
-				background-color: ${colorInfo.activityBarBackground};
-			`);
+					position: absolute;
+					width: ${layoutInfo.activityBarWidth}px;
+					height: calc(100% - ${layoutInfo.titleBarHeight + layoutInfo.statusBarHeight}px);
+					top: ${layoutInfo.titleBarHeight}px;
+					${layoutInfo.sideBarSide}: 0;
+					background-color: ${colorInfo.activityBarBackground};
+				`);
 			splash.appendChild(activityDiv);
 
 			if (colorInfo.activityBarBorder && layoutInfo.activityBarWidth > 0) {
 				const activityBorderDiv = document.createElement('div');
 				activityBorderDiv.setAttribute('style', `
-					position: absolute;
-					width: 1px;
-					height: 100%;
-					top: 0;
-					${layoutInfo.sideBarSide === 'left' ? 'right' : 'left'}: 0;
-					${layoutInfo.sideBarSide === 'left' ? 'border-right' : 'border-left'}: 1px solid ${colorInfo.activityBarBorder};
-				`);
+						position: absolute;
+						width: 1px;
+						height: 100%;
+						top: 0;
+						${layoutInfo.sideBarSide === 'left' ? 'right' : 'left'}: 0;
+						${layoutInfo.sideBarSide === 'left' ? 'border-right' : 'border-left'}: 1px solid ${colorInfo.activityBarBorder};
+					`);
 				activityDiv.appendChild(activityBorderDiv);
 			}
 
@@ -249,26 +222,26 @@
 			if (configuration.workspace) {
 				const sideDiv = document.createElement('div');
 				sideDiv.setAttribute('style', `
-					position: absolute;
-					width: ${layoutInfo.sideBarWidth}px;
-					height: calc(100% - ${layoutInfo.titleBarHeight + layoutInfo.statusBarHeight}px);
-					top: ${layoutInfo.titleBarHeight}px;
-					${layoutInfo.sideBarSide}: ${layoutInfo.activityBarWidth}px;
-					background-color: ${colorInfo.sideBarBackground};
-				`);
+						position: absolute;
+						width: ${layoutInfo.sideBarWidth}px;
+						height: calc(100% - ${layoutInfo.titleBarHeight + layoutInfo.statusBarHeight}px);
+						top: ${layoutInfo.titleBarHeight}px;
+						${layoutInfo.sideBarSide}: ${layoutInfo.activityBarWidth}px;
+						background-color: ${colorInfo.sideBarBackground};
+					`);
 				splash.appendChild(sideDiv);
 
 				if (colorInfo.sideBarBorder && layoutInfo.sideBarWidth > 0) {
 					const sideBorderDiv = document.createElement('div');
 					sideBorderDiv.setAttribute('style', `
-						position: absolute;
-						width: 1px;
-						height: 100%;
-						top: 0;
-						right: 0;
-						${layoutInfo.sideBarSide === 'left' ? 'right' : 'left'}: 0;
-						${layoutInfo.sideBarSide === 'left' ? 'border-right' : 'border-left'}: 1px solid ${colorInfo.sideBarBorder};
-					`);
+							position: absolute;
+							width: 1px;
+							height: 100%;
+							top: 0;
+							right: 0;
+							${layoutInfo.sideBarSide === 'left' ? 'right' : 'left'}: 0;
+							${layoutInfo.sideBarSide === 'left' ? 'border-right' : 'border-left'}: 1px solid ${colorInfo.sideBarBorder};
+						`);
 					sideDiv.appendChild(sideBorderDiv);
 				}
 			}
@@ -276,28 +249,28 @@
 			// part: statusbar
 			const statusDiv = document.createElement('div');
 			statusDiv.setAttribute('style', `
-				position: absolute;
-				width: 100%;
-				height: ${layoutInfo.statusBarHeight}px;
-				bottom: 0;
-				left: 0;
-				background-color: ${configuration.workspace ? colorInfo.statusBarBackground : colorInfo.statusBarNoFolderBackground};
-			`);
+					position: absolute;
+					width: 100%;
+					height: ${layoutInfo.statusBarHeight}px;
+					bottom: 0;
+					left: 0;
+					background-color: ${configuration.workspace ? colorInfo.statusBarBackground : colorInfo.statusBarNoFolderBackground};
+				`);
 			splash.appendChild(statusDiv);
 
 			if (colorInfo.statusBarBorder && layoutInfo.statusBarHeight > 0) {
 				const statusBorderDiv = document.createElement('div');
 				statusBorderDiv.setAttribute('style', `
-					position: absolute;
-					width: 100%;
-					height: 1px;
-					top: 0;
-					border-top: 1px solid ${colorInfo.statusBarBorder};
-				`);
+						position: absolute;
+						width: 100%;
+						height: 1px;
+						top: 0;
+						border-top: 1px solid ${colorInfo.statusBarBorder};
+					`);
 				statusDiv.appendChild(statusBorderDiv);
 			}
 
-			document.body.appendChild(splash);
+			window.document.body.appendChild(splash);
 		}
 
 		performance.mark('code/didShowPartsSplash');
